+++ web/public/index.js
@@ -11,8 +11,9 @@ trmnlp.connectLiveRender = function () {
     const payload = JSON.parse(msg.data);
 
     if (payload.type === "reload") {
-      trmnlp.setFrameSrc(trmnlp.iframe.src);
+      trmnlp.fetchPreview();
       trmnlp.userData.textContent = JSON.stringify(payload.user_data, null, 2);
+      hljs.highlightAll();
     }
   };
 
@@ -23,19 +24,16 @@ trmnlp.connectLiveRender = function () {
 };
 
 
-trmnlp.setPreviewFormat = function () {
-  const value = trmnlp.formatSelect.value;
-  localStorage.setItem("trmnlp-format", value);
+trmnlp.fetchPreview = function (pickerState) {
+  const screenClasses = (pickerState?.screenClasses || trmnlp.picker.state.screenClasses).join(" ");
+  const encodedScreenClasses = encodeURIComponent(screenClasses);
+  const src = `/render/${trmnlp.view}.${trmnlp.formatSelect.value}?screen_classes=${encodedScreenClasses}`;
 
-  trmnlp.setFrameSrc(`/render/${trmnlp.view}.${value}`);
-};
-
-trmnlp.setFrameSrc = function (src) {
   trmnlp.spinner.style.display = "inline-block";
   trmnlp.iframe.src = src;
 };
 
-document.addEventListener("DOMContentLoaded", function () {
+document.addEventListener("DOMContentLoaded", async function () {
   trmnlp.view = document.querySelector("meta[name='trmnl-view']").content;
   trmnlp.iframe = document.querySelector("iframe");
   trmnlp.formatSelect = document.querySelector(".select-format");
@@ -48,17 +46,24 @@ document.addEventListener("DOMContentLoaded", function () {
     trmnlp.connectLiveRender();
   }
 
-  const caseValue = localStorage.getItem("trmnlp-case") || "black";
   const formatValue = localStorage.getItem("trmnlp-format") || "html";
 
   trmnlp.formatSelect.value = formatValue;
   trmnlp.formatSelect.addEventListener("change", () => {
-    trmnlp.setPreviewFormat();
+    localStorage.setItem("trmnlp-format", trmnlp.formatSelect.value);
+    trmnlp.fetchPreview();
   });
 
   trmnlp.iframe.addEventListener("load", () => {
     trmnlp.spinner.style.display = "none";
   });
-  
-  trmnlp.setPreviewFormat();
+
+  document.getElementById('picker-form').addEventListener('trmnl:change', (event) => {
+    trmnlp.iframe.style.width = `${event.detail.width}px`;
+    trmnlp.iframe.style.height = `${event.detail.height}px`;
+
+    trmnlp.fetchPreview(event.detail);
+  });
+
+  trmnlp.picker = await TRMNLPicker.create('picker-form', { localStorageKey: 'trmnlp-picker' });
 });